pyproject.toml
[project]
name = "wizard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.103",
    "fastapi-filter>=2.0.1",
    "fastapi-pagination>=0.15.4",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "python-ulid>=3.1.0",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "blue>=0.9.1",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
